+++ ChessProject-Csharp/src/Enums/MoveResult.cs
@@ -1,10 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace SolarWinds.MSP.Chess
+﻿namespace SolarWinds.MSP.Chess.Enums
 {
     public enum MoveResult
     {
